Create base class for kickstart methods
Instead of having UEFI directly inherit of Kickstart which could fail
silently, they inherit of a shared base. This makes it easier to see
what code is shared between the methods and what is specific.
